[FM-17] :recycle: enttiy ID Long -> Integer 변경

+++ src/main/java/com/farmmate/chatroom/dto/response/ChatRoomRegistrationResponse.java
@@ -6,7 +6,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 
 @Builder(access = AccessLevel.PRIVATE)
-public record ChatRoomRegistrationResponse(Long cropId, String cropName, Long chatRoomId) {
+public record ChatRoomRegistrationResponse(Integer cropId, String cropName, Integer chatRoomId) {
 	public static ChatRoomRegistrationResponse from(ChatRoom chatRoom) {
 		return ChatRoomRegistrationResponse.builder()
 			.cropId(chatRoom.getCrop().getId())
