+++ src/main/java/com/farmmate/crop/dto/response/CropTypeResponse.java
@@ -5,7 +5,7 @@ import com.farmmate.crop.entity.Crop;
 import lombok.Builder;
 
 @Builder
-public record CropTypeResponse(Long cropId, String cropName) {
+public record CropTypeResponse(Integer cropId, String cropName) {
 	public static CropTypeResponse from(Crop crop) {
 		return CropTypeResponse.builder()
 			.cropId(crop.getId())
